Cleaned up register/otp/shadow
