+++ examples/rt685s/application/src/main.rs
@@ -18,6 +18,7 @@ use embassy_imxrt::{
 use embassy_sync::blocking_mutex::raw::NoopRawMutex;
 use embassy_time::{Duration, Instant, Timer};
 use example_bsp::application::{ExternalStorageConfig, ExternalStorageMap};
+use imxrt_rom::{otp::OtpRegisterBlock, registers::ShadowRegister};
 use partition_manager::PartitionManager;
 
 #[allow(dead_code)]
@@ -142,9 +143,9 @@ async fn main(_spawner: Spawner) {
     };
 
     {
-        let rkth_shadow = imxrt_rom::shadow::Rkth::read_shadow();
+        let rkth_shadow = imxrt_rom::registers::Rkth::read_shadow();
         let mut otp = imxrt_rom::otp::Otp::init(SYSTEM_CORE_CLOCK_HZ);
-        let rkth_otp = defmt_or_log::unwrap!(imxrt_rom::shadow::Rkth::read_fuse(&mut otp));
+        let rkth_otp = defmt_or_log::unwrap!(imxrt_rom::registers::Rkth::read_fuse(&mut otp));
 
         defmt_or_log::info!("Shadow: {:?}", rkth_shadow);
         defmt_or_log::info!("OTP:    {:?}", rkth_otp);
@@ -172,19 +173,24 @@ async fn main(_spawner: Spawner) {
             defmt_or_log::info!("USER1");
 
             if start.elapsed() > FUSE_DELAY {
-                let rkth_shadow = imxrt_rom::shadow::Rkth::read_shadow();
-                let mut otp = imxrt_rom::otp::Otp::init(SYSTEM_CORE_CLOCK_HZ);
-                let rkth_otp = defmt_or_log::unwrap!(imxrt_rom::shadow::Rkth::read_fuse(&mut otp));
+                {
+                    let rkth_shadow = imxrt_rom::registers::Rkth::read_shadow();
+                    let mut otp = imxrt_rom::otp::Otp::init(SYSTEM_CORE_CLOCK_HZ);
+                    let rkth_otp = defmt_or_log::unwrap!(imxrt_rom::registers::Rkth::read_fuse(&mut otp));
 
-                let rkth_shadow_is_default = rkth_shadow.0.as_slice().iter().all(|b| *b == 0x00);
+                    let rkth_shadow_is_default = rkth_shadow.0.as_slice().iter().all(|b| *b == 0x00);
 
-                if rkth_otp != rkth_shadow {
-                    if rkth_shadow_is_default {
-                        defmt_or_log::error!("Requesting write of fuses, but RKTH is not set to something useful");
-                    } else {
-                        defmt_or_log::info!("Writing fuses");
-                        defmt_or_log::unwrap!(rkth_shadow.write_fuse(&mut otp, false));
+                    if rkth_otp != rkth_shadow {
+                        if rkth_shadow_is_default {
+                            defmt_or_log::error!("Requesting write of fuses, but RKTH is not set to something useful");
+                        } else {
+                            defmt_or_log::info!("Writing fuses");
+                            defmt_or_log::unwrap!(rkth_shadow.write_fuse(&mut otp, false));
+                        }
                     }
+                }
+                {
+                    // TODO boot0
                 }
             } else {
                 defmt_or_log::info!("Writing new state: {}", new_state);
